python/Flask_web/app1/requirements.txt
alembic==0.9.6
amqp==2.2.2
apturl==0.5.2
arrow==0.12.0
asn1crypto==0.23.0
astroid==1.5.3
attrs==17.2.0
Automat==0.6.0
autopep8==1.3.2
Babel==2.5.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.1.2
blinker==1.4
Brlapi==0.6.5
cached-property==1.3.1
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
ChatterBot==0.7.6
chatterbot-corpus==1.0.1
click==6.7
command-not-found==0.3
constantly==15.1.0
coverage==4.4.2
cryptography==2.0.3
cssselect==1.0.1
cupshelpers==1.0
cycler==0.10.0
dateparser==0.6.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.5.0
distro==1.2.0
distro-info==0.17
Django==1.11.6
dominate==2.3.1
face-recognition-models==0.3.0
fastcache==1.0.2
feedparser==5.1.3
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Babel==0.11.2
flask-blueprint==1.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-manager==0.0.1a0
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
flask-mongoengine==0.9.3
Flask-MySQL==1.4.0
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Views==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
grequests==0.3.0
guess-language==0.2
html5lib==1.0.1
httplib2==0.10.3
hyperlink==17.3.1
idna==2.6
image==1.5.11
incremental==17.5.0
infinity==1.4
intervals==0.8.1
isort==4.2.15
itsdangerous==0.24
jieba==0.39
Jinja2==2.9.6
jsondatabase==0.1.7
keyring==10.4.0
keyrings.alt==2.2
kombu==4.1.0
language-selector==0.1
launchpadlib==1.10.5
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
louis==3.0.0
lxml==4.0.0
Mako==1.0.7
markdown2==2.3.5
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
moment==0.8.2
mongoengine==0.15.0
nltk==3.2.4
numpy==1.13.1
oauth==1.0.1
oauthlib==2.0.2
olefile==0.44
opencv==1.0.1
opencv-python==3.4.0.12
paramiko==2.4.0
parsel==1.2.0
pbr==3.1.1
pep8==1.7.1
pexpect==4.2.1
Pillow==4.2.1
psycopg2==2.7.1
py==1.4.34
pyasn1==0.3.7
pyasn1-modules==0.1.4
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pyftpdlib==1.5.3
pygame==1.9.3
pygobject==3.24.1
PyJWT==1.4.2
pylint==1.7.2
pymongo==3.5.1
PyMySQL==0.7.11
PyNaCl==1.2.1
PyNLPIR==0.5.2
pyOpenSSL==17.3.0
pyparsing==2.2.0
PyQt5==5.9
pytesseract==0.1.7
pytest==3.2.1
python-apt==1.4.0b3
python-dateutil==2.6.1
python-debian==0.1.30
python-editor==1.0.3
python-twitter==3.3
python3-openid==3.1.0
pytz==2017.2
pyxdg==0.25
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
regex==2017.12.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
ruamel.yaml==0.15.0
scikit-learn==0.19.0
Scrapy==1.4.0
SecretStorage==2.3.1
service-identity==17.0.0
simplejson==3.11.1
sip==4.19.3
six==1.11.0
sklearn==0.0
snownlp==0.12.3
SQLAlchemy==1.1.14
sqlalchemy-migrate==0.11.0
SQLAlchemy-Utils==0.32.21
sqlparse==0.2.4
suffix-trees==0.2.3
system-service==0.3
systemd-python==234
Tempita==0.5.2
textblob==0.13.0
times==0.7
tornado==4.5.2
Twisted==17.9.0
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
uWSGI==2.0.15
validators==0.12.0
vine==1.1.4
visitor==0.1.3
w3lib==1.18.0
wadllib==1.3.2
webencodings==0.5.1
Werkzeug==0.12.2
Whoosh==2.7.4
wrapt==1.10.10
WTForms==2.1
WTForms-Alchemy==0.16.5
WTForms-Components==0.10.3
xkit==0.0.0
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zope.interface==4.4.3
